Decorator nullable check

+++ src/Services/UserManagementService/src/UsersManagementService.BLL/Services/Decorators/ImagesServiceValidationDecorator.cs
@@ -2,6 +2,7 @@
 using UsersManagementService.BLL.Interfaces.Services;
 using UsersManagementService.BLL.Interfaces.Validators;
 using UsersManagementService.BLL.Models.Image;
+using UsersManagementService.Common.Exceptions;
 using static UsersManagementService.BLL.Constants.ValidationRules.ImageValidationRules;
 
 namespace UsersManagementService.BLL.Services.Decorators;
@@ -11,6 +12,11 @@ public class ImagesServiceValidationDecorator(IImagesService imagesService, IIma
     public async Task<Guid> CreateImageAsync(ImageModel image, CancellationToken cancellationToken = default)
     {
         var createImageModelValidator = imagesValidator.ImageModelValidator;
+
+        if (createImageModelValidator is null)
+        {
+            throw new NotFoundException(nameof(createImageModelValidator), null!);
+        }
 
         await createImageModelValidator.ValidateAsync(image, options =>
         {
@@ -25,6 +31,11 @@ public class ImagesServiceValidationDecorator(IImagesService imagesService, IIma
     {
         var deleteImageModelValidator = imagesValidator.ImageIdValidator;
 
+        if (deleteImageModelValidator is null)
+        {
+            throw new NotFoundException(nameof(deleteImageModelValidator), null!);
+        }
+
         await deleteImageModelValidator.ValidateAndThrowAsync(id, cancellationToken);
 
         return await imagesService.DeleteImageAsync(id, cancellationToken);
@@ -33,6 +44,11 @@ public class ImagesServiceValidationDecorator(IImagesService imagesService, IIma
     public async Task<Guid> UpdateImageAsync(ImageModel image, CancellationToken cancellationToken = default)
     {
         var updateImageModelValidator = imagesValidator.ImageModelValidator;
+
+        if (updateImageModelValidator is null)
+        {
+            throw new NotFoundException(nameof(updateImageModelValidator), null!);
+        }
 
         await updateImageModelValidator.ValidateAsync(image, options =>
         {
